{pedestrians} uncontrolled links (not part of a junction logic) now have index -1
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/pedestrians@15638 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ src/microsim/MSInternalJunction.cpp
@@ -92,10 +92,10 @@ MSInternalJunction::postloadInit() {
             myInternalLinkFoes.push_back(*j);
         }
     }
-    thisLink->setRequestInformation(requestPos, true, false, myInternalLinkFoes, myInternalLaneFoes);
+    thisLink->setRequestInformation((int)requestPos, true, false, myInternalLinkFoes, myInternalLaneFoes);
     assert(thisLink->getViaLane()->getLinkCont().size() == 1);
     MSLink* exitLink = thisLink->getViaLane()->getLinkCont()[0];
-    exitLink->setRequestInformation(requestPos, false, false, std::vector<MSLink*>(),
+    exitLink->setRequestInformation((int)requestPos, false, false, std::vector<MSLink*>(),
                                     myInternalLaneFoes, thisLink->getViaLane());
     for (std::vector<MSLink*>::const_iterator k = myInternalLinkFoes.begin(); k != myInternalLinkFoes.end(); ++k) {
         thisLink->addBlockedLink(*k);
